refactor: update coverage provider interface
