+++ packages/coverage-istanbul/src/constants.ts
@@ -1,0 +1,1 @@
+export const COVERAGE_STORE_KEY = '__VITEST_COVERAGE__'
