+++ packages/vitest/src/integrations/coverage.ts
@@ -1,5 +1,5 @@
 import { importModule } from 'local-pkg'
-import type { CoverageOptions, CoverageProvider } from '../types'
+import type { CoverageOptions, CoverageProvider, CoverageProviderModule } from '../types'
 
 export const CoverageProviderMap = {
   c8: '@vitest/coverage-c8',
@@ -8,15 +8,15 @@ export const CoverageProviderMap = {
 
 export async function getCoverageProvider(options?: CoverageOptions): Promise<CoverageProvider | null> {
   if (options?.enabled && options?.provider) {
-    const { default: CoverageProvider } = await importModule(CoverageProviderMap[options.provider])
-    return new CoverageProvider()
+    const { getProvider } = await importModule<CoverageProviderModule>(CoverageProviderMap[options.provider])
+    return await getProvider()
   }
   return null
 }
 
-export async function getCoverageInsideWorker(options: CoverageOptions) {
+export async function takeCoverageInsideWorker(options: CoverageOptions) {
   if (options.enabled && options.provider) {
-    const { default: CoverageProvider } = await importModule(CoverageProviderMap[options.provider])
-    return CoverageProvider?.getCoverage()
+    const { takeCoverage } = await importModule<CoverageProviderModule>(CoverageProviderMap[options.provider])
+    return await takeCoverage?.()
   }
 }
